+++ tests/actors/test_actor_resources.py
@@ -1,4 +1,4 @@
-from uuid import UUID
+from unittest.mock import AsyncMock, MagicMock
 
 import anyio
 import pytest
@@ -6,19 +6,8 @@ import pytest
 from jumpstarter.actors import Actor
 from jumpstarter.resources import NotAResourceError, ResourceUnavailable, resource
 from jumpstarter.states import ActorState
-from tests.mock import AsyncMock, MagicMock
 
 pytestmark = pytest.mark.anyio
-
-
-async def test_actor_id_is_a_uuid():
-    class FakeActor(Actor):
-        ...
-
-    fake_actor = FakeActor()
-
-    assert fake_actor.actor_id
-    assert isinstance(fake_actor.actor_id, UUID)
 
 
 @pytest.mark.parametrize(
@@ -204,21 +193,3 @@ async def test_resource_unavailable():
     await fake_actor.start()
 
     assert fake_actor.resource is None
-
-
-async def test_actor_can_transition_back_to_starting_after_stopped():
-    class FakeActor(Actor):
-        ...
-
-    fake_actor = FakeActor()
-
-    assert fake_actor.state == ActorState.initializing
-
-    await fake_actor.start()
-    assert fake_actor.state == ActorState.started
-
-    await fake_actor.stop()
-    assert fake_actor.state == ActorState.stopped
-
-    await fake_actor.start()
-    assert fake_actor.state == ActorState.started
